pyproject.toml
[project]
name = "rosetta-alpha"
version = "0.1.0"
description = "Dalio-inspired multi-language AI financial research platform on Arc"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Mihai Chindris", email = "chindris.mihai.alexandru@gmail.com" },
]
keywords = ["ai-agents", "trading-agents", "arc", "circle", "usdc", "prediction-markets", "risk-parity"]

dependencies = [
    # LLM orchestration (openai is a required transitive dep of adalflow)
    "adalflow>=0.2.0",
    "openai>=1.30",
    "groq>=0.9",
    # Data validation (domain models)
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    # HTTP / MCP clients
    "httpx>=0.27",
    "aiohttp>=3.9",
    # Web3 / Arc
    "web3>=6.15",
    "eth-account>=0.11",
    # API
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
    # Reasoning trace storage
    "ipfshttpclient>=0.8.0a2 ; python_version < '3.13'",
    # Regional data
    "tushare>=1.4.0",
    # Utilities
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "rich>=13.7",
    "orjson>=3.10",
    "tenacity>=8.2",
    "yfinance>=1.3.0",
    "google-genai>=2.2.0",
    "google-generativeai>=0.8.6",
    "akshare>=1.18.60",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipython>=8.20",
]
contracts = [
    # Solidity tooling is via npm/foundry — no Python deps. Placeholder.
]

[project.scripts]
rosetta = "api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "portfolio", "reasoning", "markets", "data", "workflows", "api"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
